+++ transports/dns/CHANGELOG.md
@@ -5,6 +5,8 @@
 
 [PR 3715]: https://github.com/libp2p/rust-libp2p/pull/3715
 
+[PR 3715]: https://github.com/libp2p/rust-libp2p/pull/3715
+
 ## 0.39.0
 
 - Update to `libp2p-core` `v0.39.0`.
@@ -49,6 +51,7 @@
 
 - Update to `trust-dns` `v0.21`. See [PR 2543].
 
+- make DnsConfig::custom infallible
 [PR 2543]: https://github.com/libp2p/rust-libp2p/pull/2543
 
 ## 0.32.0 [2022-02-22]